Create WEBDAV Lab commands
Microsoft IIS Webdav is a proprietary extensible web server software developed by microsoft to use with windows NT 
this protocol can be used to host websites apps and provides admins with a robust GUI for website management
IIS can be used to host both static and dynamic webpages developed in ASP.NEt and PHP The Protocols is typically configured on port 80 and port 443 The Supported Executable file extensions are:
.asp
.aspx
.config
.php 
Web Dav is a set of extensions to the http protocol which allows users o collaboratively edit and manage files on remote web servers. WebDAv essentially enables a web server to functions as a file server for collaborative authoring.WebDav run on top of Microsoft IIS ports 80/443. In order to connect to a WebDav server, we will reuqire legitamite credentials. this is becuase WebDav Implements authentication in the form of a username and password.